Operators improved.
Split now generates a truly random partition, and rejects when that partition is trivial.
Use the same data structure (Input<IntegerParameter>) for the SelectedPrior as for the other objects.
Having read Green1995, I see the need to account for the rate parameter, and added it to the Hastings Ratio.
The results, however, are still biased towards lower $k$, and SelectedPrior does not work yet, because generating a new Parameter is not possible that easily.